answer labels were not aggregated correct
Cause 1:
We used the answer_key. Every (eg gender) question has his own set of
answer possibilities. This resulted in multiple sets of labels for the
(gender) question. (see also solution 2)
cause 2
To determin the display order of the answer labels we generate a
sequence number, this is a combination of answer_number (webenq3) and an
unique id.
If you create a report with a question from multiple questionnaires, the
answer labels are sorted over this unique sequence number. This results
in multiple entries of males and females. And now come the strange bit.
If you add another question which is only available in one
questionnaire, pentaho/saiku seems to be able to create the correct
table... So I guess it is also a bug in pentaho/saiku, and I'll contact
them about it.

For now I'll apply a quick fix and remove the sorting of the answer
labels/question/...
----> so back to puc output, without ordering of results.

Future:
Solution 1
We can adjust the ETL (generation of sequence number) so this alleged
bug does not effect us. However using answer_number from webenq3 only
reduces the problem, since often female is 1 and sometimes it is 2.

Solution 2
A better approach is to adjust the ETL proces more extensive. At this
moment (eg) the gender question from one questionnaire is a different
question than the gender question from another questionnaire. This
enables the storage of the sequence within a questionnaire. But it has
the disadvantage that it is more difficult to use a specific question
for instance to create a new dimension with background characteristics.
We can also not use the keys to refer to the question because this
results in multiple 'gender question'/answer labels. Therefore the speed
of the reporting tool is lower.